+++ sktime/proba/tests/test_proba_basic.py
@@ -7,7 +7,7 @@ __author__ = ["fkiraly"]
 
 def test_proba_example():
     """Test one subsetting case for BaseDistribution."""
-    from sktime.proba.tfd import Normal
+    from sktime.proba.tfp import Normal
 
     n = Normal(mean=[[0, 1], [2, 3], [4, 5]], sd=1)
